Mongo client with no retries -- throws errors if Mongo is down.

+++ spec/spec_helper.rb
@@ -3,6 +3,8 @@ require 'fileutils'
 SPEC_DIR = File.dirname(File.expand_path(__FILE__))
 TMP_DIR  = File.join(SPEC_DIR, '..', 'tmp')
 
+trap('INT') { FileUtils.rm_rf(TMP_DIR) }
+
 require File.join(SPEC_DIR, '..', 'lib', 'weebl')
 
 RSpec.configure do |config|
@@ -15,16 +17,20 @@ module Helper
     def initialize(config)
       @config = config
       %w[left right].each { |side| FileUtils.mkdir_p(File.join(TMP_DIR, side)) }
-      
+    end
+    
+    def startup
       @procs = {
-        :left  => Daemon.new("mongod --dbpath #{TMP_DIR}/left --port #{config[:left]} --pairwith localhost:#{config[:right]}"),
-        :right => Daemon.new("mongod --dbpath #{TMP_DIR}/right --port #{config[:right]} --pairwith localhost:#{config[:left]}")
+        :left  => Daemon.new("mongod --dbpath #{TMP_DIR}/left --port #{@config[:left]} --pairwith localhost:#{@config[:right]}"),
+        :right => Daemon.new("mongod --dbpath #{TMP_DIR}/right --port #{@config[:right]} --pairwith localhost:#{@config[:left]}")
       }
     end
     
     def shutdown
+      return unless @procs
       @procs.each { |side, proc| proc.kill 'INT' }
       %w[left right].each { |side| FileUtils.rm_rf(File.join(TMP_DIR, side)) }
+      @procs = nil
     end
   end
